edited Makefile and source code of game

+++ clear.py
@@ -6,6 +6,8 @@ def delete_files_in_folder(folder_path):
     try:
         for filename in os.listdir(folder_path):
             file_path = os.path.join(folder_path, filename)
+            if "gitkeep" in file_path:
+                continue
             if os.path.isfile(file_path):
                 os.remove(file_path)
                 print(f"Deleted file: {file_path}")
